requirements.txt
# List of required packages (Dependencies)

# If you need to specify a version
requests==2.16.5
python-telegram-bot==5.0.0
pytz==2016.6.1

# Or you need not a pip package, @'...' specifies git branch
git+https://github.com/Kylmakalle/pyTelegramBotAPI.git@feature-BotAPIV3

# Or you have .zip library
https://site.org/files/package.zip
